Split up integration tests
