+++ yorm/test/test_all_files.py
@@ -1,0 +1,84 @@
+#!/usr/bin/env python
+# pylint:disable=R0201,C0111
+
+"""Integration tests for file IO."""
+
+import logging
+
+import pytest
+
+from . import refresh_file_modification_times
+from .samples import *  # pylint: disable=W0401,W0614
+
+integration = pytest.mark.integration
+
+
+@integration
+class TestInit:
+
+    """Integration tests for initializing mapped classes."""
+
+    def test_fetch_from_existing(self, tmpdir):
+        """Verify attributes are updated from an existing file."""
+        tmpdir.chdir()
+        sample = SampleStandardDecorated('sample')
+        sample2 = SampleStandardDecorated('sample')
+        assert sample2.yorm_mapper.path == sample.yorm_mapper.path
+
+        refresh_file_modification_times()
+
+        logging.info("changing values in object 1...")
+        sample.object = {'key2': 'value'}
+        sample.array = [0, 1, 2]
+        sample.string = "Hello, world!"
+        sample.number_int = 42
+        sample.number_real = 4.2
+        sample.true = True
+        sample.false = False
+
+        logging.info("reading changed values in object 2...")
+        assert 'value' == sample2.object.get('key2')
+        assert [0, 1, 2] == sample2.array
+        assert "Hello, world!" == sample2.string
+        assert 42 == sample2.number_int
+        assert 4.2 == sample2.number_real
+        assert True is sample2.true
+        assert False is sample2.false
+
+
+@integration
+class TestDelete:
+
+    """Integration tests for deleting files."""
+
+    def test_read(self, tmpdir):
+        """Verify a deleted file cannot be read from."""
+        tmpdir.chdir()
+        sample = SampleStandardDecorated('sample')
+        sample.yorm_mapper.delete()
+
+        with pytest.raises(FileNotFoundError):
+            print(sample.string)
+
+        with pytest.raises(FileNotFoundError):
+            sample.string = "def456"
+
+    def test_write(self, tmpdir):
+        """Verify a deleted file cannot be written to."""
+        tmpdir.chdir()
+        sample = SampleStandardDecorated('sample')
+        sample.yorm_mapper.delete()
+
+        with pytest.raises(FileNotFoundError):
+            sample.string = "def456"
+
+    def test_multiple(self, tmpdir):
+        """Verify a deleted file can be deleted again."""
+        tmpdir.chdir()
+        sample = SampleStandardDecorated('sample')
+        sample.yorm_mapper.delete()
+        sample.yorm_mapper.delete()
+
+
+if __name__ == '__main__':
+    pytest.main()
